requirements.txt

streamlit
openpyxl
pymupdf
openai
python-dotenv
